requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
greenlet==3.0.1
hexbytes==0.3.1
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
markup==0.2
MarkupSafe==2.1.3
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.24.4
psycopg2==2.9.9
py-solc-x==1.1.1
pycryptodome==3.19.0
PyJWT==2.8.0
pyrsistent==0.20.0
pytz==2023.3.post1
pyunormalize==15.0.0
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
semantic-version==2.10.0
SQLAlchemy==2.0.22
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.0.7
web3==6.11.3
websockets==12.0
Werkzeug==2.2.2
yarl==1.9.2
